% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_table_character_dimensions.R
\name{character_dimension}
\alias{character_dimension}
\alias{character_dimension.dimension_table}
\title{Transform a dimension numeric attributes to character}
\usage{
character_dimension(
  dimension,
  length_integers = TRUE,
  NA_replacement_value = NULL
)

\method{character_dimension}{dimension_table}(
  dimension,
  length_integers = TRUE,
  NA_replacement_value = NULL
)
}
\arguments{
\item{dimension}{A \code{dimension_table} object.}

\item{length_integers}{A \code{list} of pairs name = length, for each attribute name its length.}

\item{NA_replacement_value}{A string, value to replace NA values.}
}
\value{
A \code{dimension_table} object.
}
\description{
Transforms numeric type attributes of a dimension into character type.
}
\details{
It allows indicating the amplitude for some fields, filling with zeros on the
left: This is useful to make the alphabetical order of the result correspond
to the numerical order. It also allows indicating the literal to be used in
case the numerical value is not defined. For dates, for not defined values,
the value "9999-12-31" is assigned.
}
\keyword{internal}
