% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellation.R
\name{conform_all_dimensions}
\alias{conform_all_dimensions}
\title{Conform all dimensions of a constellation}
\usage{
conform_all_dimensions(ct)
}
\arguments{
\item{ct}{A \code{constellation} object.}
}
\value{
A \code{constellation} object.
}
\description{
Conform all dimensions with the same name in the star schemas of a
constellation. If two dimensions have the same name in a constellation, they
must be conformed.
}
\keyword{internal}
