% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellation.R
\name{conform_dimensions}
\alias{conform_dimensions}
\title{Conform dimensions of given name}
\usage{
conform_dimensions(ct, name = NULL)
}
\arguments{
\item{ct}{A \code{constellation} object.}

\item{name}{A string, name of the dimension.}
}
\value{
A \code{constellation} object.
}
\description{
If two dimensions have the same name in a constellation, they must be
conformed.
}
\keyword{internal}
