% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellation.R
\name{constellation}
\alias{constellation}
\title{\code{constellation} S3 class}
\usage{
constellation(lst, name = NULL)
}
\arguments{
\item{lst}{A list of \code{star_schema} objects.}

\item{name}{A string.}
}
\value{
A \code{constellation} object.
}
\description{
Creates a \code{constellation} object from a list of \code{star_schema} objects. All
dimensions with the same name in the star schemas have to be conformable.
}
\examples{

ct <- constellation(list(st_mrs_age, st_mrs_cause), name = "mrs")

}
\seealso{


Other star schema and constellation definition functions: 
\code{\link{character_dimensions}()},
\code{\link{role_playing_dimension}()},
\code{\link{snake_case}()},
\code{\link{star_schema}()}
}
\concept{star schema and constellation definition functions}
