% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellation_as_multistar.R
\name{constellation_as_multistar}
\alias{constellation_as_multistar}
\alias{constellation_as_multistar.constellation}
\title{Export a constellation as a \code{multistar}}
\usage{
constellation_as_multistar(ct)

\method{constellation_as_multistar}{constellation}(ct)
}
\arguments{
\item{ct}{A \code{constellation} object.}
}
\value{
A \code{multistar} object.
}
\description{
Once we have refined the format or content of facts and dimensions, we can
obtain a \code{multistar}. A \code{multistar} only distinguishes between general and
conformed dimensions, each dimension has its own data. It can contain
multiple fact tables.
}
\examples{
library(tidyr)

ms <- ct_mrs \%>\%
  constellation_as_multistar()

}
\seealso{


Other results export functions: 
\code{\link{constellation_as_tibble_list}()},
\code{\link{multistar_as_flat_table}()},
\code{\link{star_schema_as_flat_table}()},
\code{\link{star_schema_as_multistar}()},
\code{\link{star_schema_as_tibble_list}()}
}
\concept{results export functions}
