% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ct_mrs}
\alias{ct_mrs}
\title{Constellation for Mortality Reporting System}
\format{
A \code{constellation} object.
}
\usage{
ct_mrs
}
\description{
Constellation for the Mortality Reporting System considering age and cause
classification.
}
\examples{
# Defined by:

ct_mrs <- constellation(list(st_mrs_age, st_mrs_cause), name = "mrs")

}
\keyword{datasets}
