% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ct_mrs_test}
\alias{ct_mrs_test}
\title{Constellation for Mortality Reporting System Test}
\format{
A \code{constellation} object.
}
\usage{
ct_mrs_test
}
\description{
Constellation for the Mortality Reporting System considering age and cause
classification data test.
}
\examples{
# Defined by:

ct_mrs_test <-
  constellation(list(st_mrs_age_test, st_mrs_cause_test), name = "mrs_test")

}
\keyword{datasets}
