% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomultistar_define_geoattribute.R
\name{define_geoattribute}
\alias{define_geoattribute}
\alias{define_geoattribute.geomultistar}
\title{Define geographic attribute}
\usage{
define_geoattribute(
  gms,
  dimension = NULL,
  attribute = NULL,
  from_layer = NULL,
  by = NULL,
  from_attribute = NULL
)

\method{define_geoattribute}{geomultistar}(
  gms,
  dimension = NULL,
  attribute = NULL,
  from_layer = NULL,
  by = NULL,
  from_attribute = NULL
)
}
\arguments{
\item{gms}{A \code{geomultistar} object.}

\item{dimension}{A string, dimension name.}

\item{attribute}{A string, attribute name.}

\item{from_layer}{A \code{sf} object.}

\item{by}{a vector of correspondence of attributes of the dimension with the
\code{sf} layer structure.}

\item{from_attribute}{A string, attribute name.}
}
\value{
A \code{geomultistar} object.
}
\description{
Associates the instances of an attribute of the geographic dimension with the
instances of a geographic layer or defines it from the geometry of another
previously defined geographic attribute.
}
\details{
If defined from a layer, additionally the attributes used for the join
between the tables must be indicated.

If defined from another attribute, it should have a finer granularity, to
obtain the result by grouping its instances.
}
\examples{
library(tidyr)
library(sf) # It has to be included even if it is not used directly.

gms <- geomultistar(ms = ms_mrs, geodimension = "where") \%>\%
  define_geoattribute(
    attribute = "city",
    from_layer = usa_cities,
    by = c("city" = "city", "state" = "state")
  ) \%>\%
  define_geoattribute(attribute = "region",
                      from_attribute = "city")

# 1
gms <- gms \%>\%
  define_geoattribute(attribute = "all_where",
                      from_attribute = "region")
# 2
gms <- gms \%>\%
  define_geoattribute(attribute = "all_where",
                      from_attribute = "city")
# 3
gms <- gms \%>\%
  define_geoattribute(attribute = "all_where",
                      from_layer = usa_nation)


}
\seealso{


Other geo functions: 
\code{\link{geomultistar}()},
\code{\link{get_empty_geoinstances}()}
}
\concept{geo functions}
