% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_query_run_query.R
\name{define_selected_dimensions}
\alias{define_selected_dimensions}
\title{Define selected dimensions}
\usage{
define_selected_dimensions(dq)
}
\arguments{
\item{dq}{A \code{dimensional_query} object.}
}
\value{
A \code{dimensional_query} object.
}
\description{
Include the selected dimensions and only the selected attributes in them.
}
\keyword{internal}
