% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_query_run_query.R
\name{define_selected_facts}
\alias{define_selected_facts}
\title{Define selected facts}
\usage{
define_selected_facts(dq)
}
\arguments{
\item{dq}{A \code{dimensional_query} object.}
}
\value{
A \code{dimensional_query} object.
}
\description{
Measure names are stored as the names of the columns with the aggregation
functions.
}
\keyword{internal}
