% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact_table_incremental_refresh.R
\name{delete_records}
\alias{delete_records}
\title{Delete records}
\usage{
delete_records(ft, ft_new, fk)
}
\arguments{
\item{ft}{A \code{fact_table} object.}

\item{ft_new}{A \code{fact_table} object.}

\item{fk}{A vector of foreign key names.}
}
\value{
A \code{fact_table} object.
}
\description{
Delete records with the same primary key.
}
\keyword{internal}
