% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_query_run_query.R
\name{delete_unused_foreign_keys}
\alias{delete_unused_foreign_keys}
\title{Delete unused foreign keys}
\usage{
delete_unused_foreign_keys(dq)
}
\arguments{
\item{dq}{A \code{dimensional_query} object.}
}
\value{
A \code{dimensional_query} object.
}
\description{
In facts, remove foreign keys from dimensions not included in the result.
}
\keyword{internal}
