% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomultistar.R
\name{dereference_dimension}
\alias{dereference_dimension}
\title{Dereference a dimension}
\usage{
dereference_dimension(ft, dimension, conversion = TRUE)
}
\arguments{
\item{ft}{A \code{fact_table} object.}

\item{dimension}{A \code{dimension_table} object.}

\item{conversion}{A boolean, indicates whether the attributes need to be
transformed.}
}
\value{
A \code{fact_table} object.
}
\description{
Given a dimension, transform the fact table so that the primary key of the
dimension (which is a foreign key in the fact table) is replaced by the other
attributes of the dimension.
}
\keyword{internal}
