% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_model.R
\name{dimensional_model}
\alias{dimensional_model}
\title{\code{dimensional_model} S3 class}
\usage{
dimensional_model()
}
\value{
A \code{dimensional_model} object.
}
\description{
An empty \code{dimensional_model} object is created in which definition of facts
and dimensions can be added.
}
\details{
To get a star schema (a \code{star_schema} object) we need a flat table
(implemented through a \code{tibble}) and a \code{dimensional_model} object. The
definition of facts and dimensions in the \code{dimensional_model} object is made
from the flat table columns. Each attribute can only appear once in the
definition.
}
\examples{

dm <- dimensional_model()

}
\seealso{
\code{\link{star_schema}}

Other star definition functions: 
\code{\link{define_dimension}()},
\code{\link{define_fact}()}
}
\concept{star definition functions}
