% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_query.R
\name{dimensional_query}
\alias{dimensional_query}
\title{\code{dimensional_query} S3 class}
\usage{
dimensional_query(ms = NULL)
}
\arguments{
\item{ms}{A \code{multistar} object.}
}
\value{
A \code{dimensional_query} object.
}
\description{
An empty \code{dimensional_query} object is created where you can select fact
measures, dimension attributes and filter dimension rows.
}
\examples{
library(tidyr)

ms_mrs <- ct_mrs \%>\%
  constellation_as_multistar()

dq <- dimensional_query(ms_mrs)

}
\seealso{


Other query functions: 
\code{\link{filter_dimension}()},
\code{\link{run_query}()},
\code{\link{select_dimension}()},
\code{\link{select_fact}()}
}
\concept{query functions}
