% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dm_mrs_age}
\alias{dm_mrs_age}
\title{Star Definition for Mortality Reporting System by Age}
\format{
A \code{dimensional_model} object.
}
\usage{
dm_mrs_age
}
\description{
Definition of facts and dimensions for the Mortality Reporting System
considering the age classification.
}
\examples{
# Defined by:

library(tidyr)

dm_mrs_age <- dimensional_model() \%>\%
  define_fact(
    name = "mrs_age",
    measures = c(
      "Deaths"
    ),
    agg_functions = c(
      "SUM"
    ),
    nrow_agg = "nrow_agg"
  ) \%>\%
  define_dimension(
    name = "when",
    attributes = c(
      "Week Ending Date",
      "WEEK",
      "Year"
    )
  ) \%>\%
  define_dimension(
    name = "when_available",
    attributes = c(
      "Data Availability Date",
      "Data Availability Week",
      "Data Availability Year"
    )
  ) \%>\%
  define_dimension(
    name = "where",
    attributes = c(
      "REGION",
      "State",
      "City"
    )
  ) \%>\%
  define_dimension(
    name = "who",
    attributes = c(
      "Age Range"
    )
  )

}
\keyword{datasets}
