% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_query_run_geoquery.R
\name{filter_geodimension}
\alias{filter_geodimension}
\title{Filter geodimension}
\usage{
filter_geodimension(dq)
}
\arguments{
\item{dq}{A \code{dimensional_query} object.}
}
\value{
A \code{dimensional_query} object.
}
\description{
Filter the geodimension instances according to the query definition.
}
\keyword{internal}
