% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_query_run_query.R
\name{filter_selected_instances}
\alias{filter_selected_instances}
\title{Filter selected instances}
\usage{
filter_selected_instances(dq)
}
\arguments{
\item{dq}{A \code{dimensional_query} object.}
}
\value{
A \code{dimensional_query} object.
}
\description{
For some dimensions the instances to include have been defined, we have the
value of the primary key. They are filtered for both facts and dimensions.
}
\keyword{internal}
