% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_table_update_dimension.R
\name{find_values}
\alias{find_values}
\title{Find values in a dimension}
\usage{
find_values(dimension, values)
}
\arguments{
\item{dimension}{A \code{dimension_table} object.}

\item{values}{A vector of named values.}
}
\value{
A vector of boolean.
}
\description{
Find a vector of named values in a dimension.
}
\keyword{internal}
