% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ft_datagov_uk}
\alias{ft_datagov_uk}
\title{Modelling the long-term health impacts of air pollution in London}
\format{
A \code{tibble}.
}
\source{
\url{https://data.world/datagov-uk/fd864906-8456-46a8-9a01-0dcb2dbd87b9}
}
\usage{
ft_datagov_uk
}
\description{
Estimation of the long-term health impacts of exposure to air pollution in
London from 2016 to 2050.
}
\details{
The original dataset contains 68 files, corresponding to 34 London areas and
2 pollutants: pollutant and zone are indicated in the name of each file. Each
file has several sheets with different variables. It has been transformed
into a flat table considering a single variable and defining the area and the
pollutant as columns.
}
\keyword{datasets}
