% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ft_london_boroughs}
\alias{ft_london_boroughs}
\title{London Boroughs}
\format{
A \code{tibble}.
}
\source{
\url{https://data.london.gov.uk/dataset/london-borough-profiles} and
\url{https://en.wikipedia.org/wiki/List_of_sub-regions_used_in_the_London_Plan}
}
\usage{
ft_london_boroughs
}
\description{
Classification of London's boroughs into zones and sub-regions.
}
\keyword{datasets}
