% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ft_usa_city_county}
\alias{ft_usa_city_county}
\title{USA City and County}
\format{
A \code{tibble}.
}
\source{
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/carto-boundary-file.html}
}
\usage{
ft_usa_city_county
}
\description{
City, state and county for US cities. It only includes those that appear in
the Mortality Reporting System.
}
\keyword{datasets}
