% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ft_usa_city_county_mrs}
\alias{ft_usa_city_county_mrs}
\title{USA City and County MRS}
\format{
A \code{tibble}.
}
\source{
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/carto-boundary-file.html}
}
\usage{
ft_usa_city_county_mrs
}
\description{
City and county for cities considered in the Mortality Reporting System.
}
\keyword{datasets}
