% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_schema_modify_dimension_records.R
\name{get_all_dimensions}
\alias{get_all_dimensions}
\title{Get all dimensions}
\usage{
get_all_dimensions(st)
}
\arguments{
\item{st}{A \code{star_schema} object.}
}
\value{
A list of \code{dimension_table} objects.
}
\description{
Get all the dimensions of a star schema.
}
\keyword{internal}
