% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_model.R
\name{get_attribute_names}
\alias{get_attribute_names}
\title{Get attribute names}
\usage{
get_attribute_names(dm)
}
\arguments{
\item{dm}{A \code{dimensional_model} object.}
}
\value{
A vector of attribute names.
}
\description{
Get the names of the attributes used so far in the definition.
}
\keyword{internal}
