% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellation_get_conformed_dimension.R
\name{get_conformed_dimension}
\alias{get_conformed_dimension}
\alias{get_conformed_dimension.constellation}
\title{Get conformed dimension}
\usage{
get_conformed_dimension(ct, name)

\method{get_conformed_dimension}{constellation}(ct, name)
}
\arguments{
\item{ct}{A \code{constellation} object.}

\item{name}{A string, name of the dimension.}
}
\value{
A \code{dimension_table} object.
}
\description{
Get a conformed dimension of a constellation given its name.
}
\examples{
library(tidyr)

d <- ct_mrs \%>\%
  get_conformed_dimension("when")

}
\seealso{
Other data cleaning functions: 
\code{\link{get_conformed_dimension_names}()},
\code{\link{get_dimension_names}()},
\code{\link{get_dimension}()},
\code{\link{match_records}()},
\code{\link{modify_conformed_dimension_records}()},
\code{\link{modify_dimension_records}()},
\code{\link{record_update_set}()},
\code{\link{update_record}()},
\code{\link{update_selection_general}()},
\code{\link{update_selection}()}
}
\concept{data cleaning functions}
