% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellation_get_conformed_dimension.R
\name{get_conformed_dimension_names}
\alias{get_conformed_dimension_names}
\alias{get_conformed_dimension_names.constellation}
\title{Get conformed dimension names}
\usage{
get_conformed_dimension_names(ct)

\method{get_conformed_dimension_names}{constellation}(ct)
}
\arguments{
\item{ct}{A \code{constellation} object.}
}
\value{
A vector of dimension names.
}
\description{
Get the names of the conformed dimensions of a constellation.
}
\examples{
library(tidyr)

d <- ct_mrs \%>\%
  get_conformed_dimension_names()

}
\seealso{


Other data cleaning functions: 
\code{\link{get_conformed_dimension}()},
\code{\link{get_dimension_names}()},
\code{\link{get_dimension}()},
\code{\link{match_records}()},
\code{\link{modify_conformed_dimension_records}()},
\code{\link{modify_dimension_records}()},
\code{\link{record_update_set}()},
\code{\link{update_record}()},
\code{\link{update_selection_general}()},
\code{\link{update_selection}()}
}
\concept{data cleaning functions}
