% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_schema_attributes.R
\name{get_conformed_dimension_names_st}
\alias{get_conformed_dimension_names_st}
\alias{get_conformed_dimension_names_st.star_schema}
\title{Get conformed dimension names}
\usage{
get_conformed_dimension_names_st(st)

\method{get_conformed_dimension_names_st}{star_schema}(st)
}
\arguments{
\item{st}{A \code{star_schema} object.}
}
\value{
A vector of dimension names.
}
\description{
Get the names of the star schema conformed dimensions.
}
\keyword{internal}
