% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_schema_get_dimension.R
\name{get_dimension_names}
\alias{get_dimension_names}
\alias{get_dimension_names.star_schema}
\title{Get dimension names}
\usage{
get_dimension_names(st)

\method{get_dimension_names}{star_schema}(st)
}
\arguments{
\item{st}{A \code{star_schema} object.}
}
\value{
A vector of dimension names.
}
\description{
Get the names of the dimensions of a star schema.
}
\details{
Role playing dimensions are not considered.
}
\examples{

dn <- st_mrs_age |>
  get_dimension_names()

}
\seealso{
Other data cleaning functions: 
\code{\link{get_conformed_dimension}()},
\code{\link{get_conformed_dimension_names}()},
\code{\link{get_dimension}()},
\code{\link{match_records}()},
\code{\link{modify_conformed_dimension_records}()},
\code{\link{modify_dimension_records}()},
\code{\link{record_update_set}()},
\code{\link{update_record}()},
\code{\link{update_selection}()},
\code{\link{update_selection_general}()}
}
\concept{data cleaning functions}
