% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_table_attributes.R
\name{get_dimension_type}
\alias{get_dimension_type}
\alias{get_dimension_type.dimension_table}
\title{Get the dimension type}
\usage{
get_dimension_type(dimension)

\method{get_dimension_type}{dimension_table}(dimension)
}
\arguments{
\item{dimension}{A \code{dimension_table} object.}
}
\value{
A string, type of the dimension.
}
\description{
Returns the type of the dimension.
}
\details{
Attributes can be accessed directly but this function has been defined
because it is used from other classes and is thus done in a more controlled
way.
}
\keyword{internal}
