% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_schema_attributes.R
\name{get_fact_name}
\alias{get_fact_name}
\alias{get_fact_name.star_schema}
\title{Get fact name}
\usage{
get_fact_name(st)

\method{get_fact_name}{star_schema}(st)
}
\arguments{
\item{st}{A \code{star_schema} object.}
}
\value{
A string, name of the fact table.
}
\description{
Get the name of the fact table.
}
\keyword{internal}
