% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_schema_rename.R
\name{get_measure_names}
\alias{get_measure_names}
\alias{get_measure_names.star_schema}
\title{Get measure names}
\usage{
get_measure_names(st)

\method{get_measure_names}{star_schema}(st)
}
\arguments{
\item{st}{A \code{star_schema} object.}
}
\value{
A vector of measure names.
}
\description{
Get the name of measures in facts.
}
\examples{
library(tidyr)

measure_names <-
  st_mrs_age \%>\% get_measure_names()

}
\seealso{


Other rename functions: 
\code{\link{get_dimension_attribute_names}()},
\code{\link{rename_dimension_attributes}()},
\code{\link{rename_dimension}()},
\code{\link{rename_fact}()},
\code{\link{rename_measures}()}
}
\concept{rename functions}
