% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_schema_as_tibble_list.R
\name{get_name_of_role_playing_dimensions}
\alias{get_name_of_role_playing_dimensions}
\title{Get the name of the role playing dimensions}
\usage{
get_name_of_role_playing_dimensions(st)
}
\arguments{
\item{st}{A \code{star_schema} object.}
}
\value{
A vector of dimension names.
}
\description{
Get the name of the role playing dimensions
}
\keyword{internal}
