% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_schema_incremental_refresh.R
\name{get_name_of_uniquely_implemented_dimensions}
\alias{get_name_of_uniquely_implemented_dimensions}
\title{Get name of uniquely implemented dimensions}
\usage{
get_name_of_uniquely_implemented_dimensions(st)
}
\arguments{
\item{st}{A \code{star_schema} object.}
}
\value{
A vector of dimension names.
}
\description{
Get a list of dimension names that are uniquely implemented.
}
\details{
For role dimensions that share role playing dimension, only one is
considered. Role playing dimensions are not considered.
}
\keyword{internal}
