% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_schema_purge_dimensions.R
\name{get_role_dimension_names}
\alias{get_role_dimension_names}
\title{Get role dimension names associated to a role-playing dimension}
\usage{
get_role_dimension_names(st, name)
}
\arguments{
\item{dimension}{A \code{star_schema} object.}
}
\value{
A vector of dimension names.
}
\description{
Each role dimension has the name of the role-playing dimension associated.
This function allows us to obtain role dimension names for a role-playing
dimension.
}
\keyword{internal}
