% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_table_attributes.R
\name{get_role_playing_dimension_name}
\alias{get_role_playing_dimension_name}
\alias{get_role_playing_dimension_name.dimension_table}
\title{Get the associated role-playing dimension name}
\usage{
get_role_playing_dimension_name(dimension)

\method{get_role_playing_dimension_name}{dimension_table}(dimension)
}
\arguments{
\item{dimension}{A \code{dimension_table} object.}
}
\value{
A string, name of the dimension.
}
\description{
Each role dimension has the name of the role-playing dimension associated.
This function allows us to obtain its name.
}
\details{
Attributes can be accessed directly but this function has been defined
because it is used from other classes and is thus done in a more controlled
way.
}
\keyword{internal}
