% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellation_get_star_schema.R
\name{get_star_schema}
\alias{get_star_schema}
\alias{get_star_schema.constellation}
\title{Get star schema}
\usage{
get_star_schema(ct, name)

\method{get_star_schema}{constellation}(ct, name)
}
\arguments{
\item{ct}{A \code{constellation} object.}

\item{name}{A string, name of the star schema.}
}
\value{
A \code{dimension_table} object.
}
\description{
Get a star schema of a constellation given its name.
}
\examples{
library(tidyr)

d <- ct_mrs \%>\%
  get_star_schema("mrs_age")

}
\seealso{


Other incremental refresh functions: 
\code{\link{filter_fact_rows}()},
\code{\link{get_star_schema_names}()},
\code{\link{incremental_refresh_constellation}()},
\code{\link{incremental_refresh_star_schema}()},
\code{\link{purge_dimensions_constellation}()},
\code{\link{purge_dimensions_star_schema}()}
}
\concept{incremental refresh functions}
