% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellation_get_star_schema.R
\name{get_star_schema_names}
\alias{get_star_schema_names}
\alias{get_star_schema_names.constellation}
\title{Get star schema names}
\usage{
get_star_schema_names(ct)

\method{get_star_schema_names}{constellation}(ct)
}
\arguments{
\item{ct}{A \code{constellation} object.}
}
\value{
A vector of star schema names.
}
\description{
Get the names of the star schemas in a constellation.
}
\examples{
library(tidyr)

d <- ct_mrs \%>\%
  get_star_schema_names()

}
\seealso{
Other incremental refresh functions: 
\code{\link{filter_fact_rows}()},
\code{\link{get_star_schema}()},
\code{\link{incremental_refresh_constellation}()},
\code{\link{incremental_refresh_star_schema}()},
\code{\link{purge_dimensions_constellation}()},
\code{\link{purge_dimensions_star_schema}()}
}
\concept{incremental refresh functions}
