% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomultistar_define_geoattribute.R
\name{\%>\%}
\alias{\%>\%}
\alias{define_geoattribute_from_attribute}
\alias{define_geoattribute_from_layer}
\title{Define a geoattribute from another}
\usage{
define_geoattribute_from_attribute(
  gms,
  dimension = NULL,
  attribute = NULL,
  from_attribute = NULL
)

define_geoattribute_from_layer(
  gms,
  dimension = NULL,
  attribute = NULL,
  from_layer = NULL,
  by = NULL
)
}
\arguments{
\item{gms}{A \code{geomultistar} object.}

\item{dimension}{A string, dimension name.}

\item{attribute}{A string, attribute name.}

\item{from_attribute}{A string, attribute name.}

\item{by}{a vector of correspondence of attributes of the dimension with the
\code{sf} structure.}

\item{layer}{A \code{sf} object}
}
\value{
A \code{geomultistar} object.

A \code{geomultistar} object.
}
\description{
Define a geoattribute from another.

Define an attribute from a layer.
}
\keyword{internal}
