% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_query_run_query.R
\name{group_facts}
\alias{group_facts}
\title{Group facts}
\usage{
group_facts(dq)
}
\arguments{
\item{dq}{A \code{dimensional_query} object.}
}
\value{
A \code{dimensional_query} object.
}
\description{
Once the external keys have been possibly replaced, group the rows of facts.
}
\keyword{internal}
