% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomultistar.R
\name{group_table}
\alias{group_table}
\title{Group the records in the table}
\usage{
group_table(ft)
}
\arguments{
\item{ft}{A \code{fact_table} object.}
}
\value{
A \code{fact_table} object.
}
\description{
Group the records in the table using the aggregation functions for the
measurements.
}
\keyword{internal}
