% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_table_homogenize.R
\name{homogenize}
\alias{homogenize}
\alias{homogenize.dimension_table}
\title{Homogenize a dimension}
\usage{
homogenize(dimension, attributes = NULL)

\method{homogenize}{dimension_table}(dimension, attributes = NULL)
}
\arguments{
\item{dimension}{A \code{dimension_table} object.}

\item{attributes}{A vector of attribute names of the dimension.}
}
\value{
A \code{dimension_table} object.
}
\description{
To merge dimensions, they must first be homogenized: the generated primary
key must be removed and, if necessary, its attributes (columns) must be
renamed.
}
\keyword{internal}
