% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_table_incremental_refresh.R
\name{incremental_refresh_dimension}
\alias{incremental_refresh_dimension}
\alias{incremental_refresh_dimension.dimension_table}
\title{Incrementally refresh a dimension with another}
\usage{
incremental_refresh_dimension(dimension, dimension_new)

\method{incremental_refresh_dimension}{dimension_table}(dimension, dimension_new)
}
\arguments{
\item{dimension}{A \code{dimension_table} object.}

\item{dimension_new}{A \code{dimension_table} object, possibly with new data.}
}
\value{
A \code{dimension_table} object.
}
\description{
Incrementally refresh a dimension with the content of a new one that is
integrated into the first.
}
\keyword{internal}
