% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_table_attributes.R
\name{is_conformed_dimension}
\alias{is_conformed_dimension}
\alias{is_conformed_dimension.dimension_table}
\title{Is it conformed dimension?}
\usage{
is_conformed_dimension(dimension)

\method{is_conformed_dimension}{dimension_table}(dimension)
}
\arguments{
\item{dimension}{A \code{dimension_table} object.}
}
\value{
A boolean.
}
\description{
Indicates by means of a boolean if the dimension is a conformed dimension.
}
\details{
Attributes can be accessed directly but this function has been defined
because it is used from other classes and is thus done in a more controlled
way.
}
\keyword{internal}
