% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_update_set_is_dimension_in_updates.R
\name{is_dimension_in_updates}
\alias{is_dimension_in_updates}
\alias{is_dimension_in_updates.record_update_set}
\title{Is dimension in set of updates?}
\usage{
is_dimension_in_updates(updates, name)

\method{is_dimension_in_updates}{record_update_set}(updates, name)
}
\arguments{
\item{updates}{A \code{record_update_set} object, list of dimension record update
operations.}

\item{name}{A string, name of the dimension.}
}
\value{
A boolean, indicating if the dimension appears in the list of update
operations.
}
\description{
Given a set of dimension record update operations and the name of a
dimension, it checks if there is any update operation to perform on the
dimension.
}
\keyword{internal}
