% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mrs_age_test}
\alias{mrs_age_test}
\title{Mortality Reporting System by Age Test}
\format{
A \code{tibble}.
}
\source{
\url{https://catalog.data.gov/dataset/deaths-in-122-u-s-cities-1962-2016-122-cities-mortality-reporting-system}
}
\usage{
mrs_age_test
}
\description{
Selection of data from the 2 Cities Mortality Reporting System by age
group, for the first 3 weeks of 1962.
}
\details{
The original dataset begins in 1962. For each week, in 122 US cities,
mortality figures by age group and cause, considered separately, are included
(i.e., the combination of age group and cause is not included). In the cause,
only a distinction is made between pneumonia or influenza and others.

Two additional dates have been generated, which were not present in the
original dataset.
}
\keyword{datasets}
