% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mrs_age_w10}
\alias{mrs_age_w10}
\title{Mortality Reporting System by Age for Week 10}
\format{
A \code{tibble}.
}
\source{
\url{https://catalog.data.gov/dataset/deaths-in-122-u-s-cities-1962-2016-122-cities-mortality-reporting-system}
}
\usage{
mrs_age_w10
}
\description{
Selection of data from the 122 Cities Mortality Reporting System by age
group, for week 10 of 1962. It also includes some isolated data from previous
weeks that is supposed to be corrections for data errors.
}
\details{
The original dataset begins in 1962. For each week, in 122 US cities,
mortality figures by age group and cause, considered separately, are included
(i.e., the combination of age group and cause is not included). In the cause,
only a distinction is made between pneumonia or influenza and others.

Two additional dates have been generated, which were not present in the
original dataset.
}
\keyword{datasets}
