% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ms_where}
\alias{ms_where}
\title{Multistar for Mortality Reporting System with the Enriched \code{where} Dimension}
\format{
A \code{multistar} object.
}
\source{
\url{https://cran.r-project.org/package=starschemar}
}
\usage{
ms_where
}
\description{
Multistar for the Mortality Reporting System considering age and cause
classification with the enriched \code{where} dimension.
}
\examples{
# Defined by:

library(tidyr)
library(starschemar)

where <- st_mrs_age \%>\%
  get_dimension("where")

updates <- record_update_set() \%>\%
  update_selection(
    dimension = where,
    columns = c("city"),
    old_values = c("Bridgepor"),
    new_values = c("Bridgeport")
  ) \%>\%
  update_selection(
    dimension = where,
    columns = c("city"),
    old_values = c("Wilimington"),
    new_values = c("Wilmington")
  )

st_mrs_age <-
  st_mrs_age \%>\%
  modify_dimension_records(updates)

st_mrs_cause <-
  st_mrs_cause \%>\%
  modify_dimension_records(updates)


st_mrs_age <-
  st_mrs_age \%>\%
  rename_dimension_attributes(
    name = "where",
    attributes = c("region"),
    new_names = c("division")
  )

st_mrs_cause <-
  st_mrs_cause \%>\%
  rename_dimension_attributes(
    name = "where",
    attributes = c("region"),
    new_names = c("division")
  )

tb <-
  enrich_dimension_export(st_mrs_age,
                          name = "where",
                          attributes = c("division"))

division_name <- c(
  "New England",
  "Middle Atlantic",
  "East North Central",
  "West North Central",
  "South Atlantic",
  "East South Central",
  "West South Central",
  "Mountain",
  "Pacific"
)

region_name <- c(
  "Northeast",
  "Northeast",
  "Midwest",
  "Midwest",
  "South",
  "South",
  "South",
  "West",
  "West"
)

region <- c('1',
            '1',
            '2',
            '2',
            '3',
            '3',
            '3',
            '4',
            '4')

tb <-
  tibble::add_column(tb,
                     division_name = division_name,
                     region = region,
                     region_name = region_name)

st_mrs_age <-
  st_mrs_age \%>\%
  enrich_dimension_import(name = "where", tb)

st_mrs_cause <-
  st_mrs_cause \%>\%
  enrich_dimension_import(name = "where", tb)

tb <- ft_usa_states \%>\%
   tibble::add_row(state = "Unknown", state_name = "Unknown")

st_mrs_age <-
  st_mrs_age \%>\%
  enrich_dimension_import(name = "where", tb)

st_mrs_cause <-
  st_mrs_cause \%>\%
  enrich_dimension_import(name = "where", tb)

tb <-
  enrich_dimension_export(st_mrs_age,
                          name = "where",
                          attributes = c("city", "state"))

tb <- ft_usa_city_county_mrs \%>\%
  tibble::add_row(city = "Unknown",
                  state = "Unknown",
                  county = "Unknown")

st_mrs_age <-
  st_mrs_age \%>\%
  enrich_dimension_import(name = "where", tb)

st_mrs_cause <-
  st_mrs_cause \%>\%
  enrich_dimension_import(name = "where", tb)

ms_where <- constellation(list(st_mrs_age, st_mrs_cause), name = "mrs") \%>\%
  constellation_as_multistar()

}
\keyword{datasets}
