% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_table.R
\name{new_dimension_table}
\alias{new_dimension_table}
\title{\code{dimension_table} S3 class}
\usage{
new_dimension_table(ft = tibble::tibble(), name = NULL, type = "general")
}
\arguments{
\item{ft}{A \code{tibble}, contains a dimension.}

\item{name}{A string, name of the dimension.}

\item{type}{A string, type of the dimension.}
}
\value{
A \code{dimension_table} object.
}
\description{
Internal low-level constructor that creates new objects with the correct
structure.
}
\details{
Dimension types:
http://dwhlaureate.blogspot.com/2013/08/types-of-dimensions.html

Types considered: (general), (role, role_playing), (conformed).
}
\keyword{internal}
