% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_model.R
\name{new_dimensional_model}
\alias{new_dimensional_model}
\title{dimensional_model S3 class}
\usage{
new_dimensional_model()
}
\value{
A \code{dimensional_model} object.
}
\description{
Internal low-level constructor that creates new objects with the correct
structure.
}
\keyword{internal}
