% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_query.R
\name{new_dimensional_query}
\alias{new_dimensional_query}
\title{\code{dimensional_query} S3 class}
\usage{
new_dimensional_query(ms = NULL)
}
\value{
A \code{dimensional_query} object.
}
\description{
Internal low-level constructor that creates new objects with the correct
structure.
}
\keyword{internal}
