% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact_table.R
\name{new_fact_table}
\alias{new_fact_table}
\title{\code{fact_table} S3 class}
\usage{
new_fact_table(
  ft = tibble::tibble(),
  name = NULL,
  measures = NULL,
  agg_functions = NULL,
  nrow_agg = NULL
)
}
\arguments{
\item{ft}{A \code{tibble}, contains the fact table.}

\item{name}{A string, name of the fact.}

\item{measures}{A vector of measurement names.}

\item{agg_functions}{A vector of aggregation function names.}

\item{nrow_agg}{A string, measurement name for the number of rows aggregated.}
}
\value{
A \code{fact_table} object.
}
\description{
Internal low-level constructor that creates new objects with the correct
structure.
}
\keyword{internal}
