% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomultistar.R
\name{new_geomultistar}
\alias{new_geomultistar}
\title{\code{geomultistar} S3 class}
\usage{
new_geomultistar(ms = NULL, geodimension = NULL)
}
\arguments{
\item{ms}{A \code{multistar} structure.}

\item{geodimension}{A vector of dimension names.}
}
\value{
A \code{geomultistar} object.
}
\description{
Internal low-level constructor that creates new objects with the correct
structure.
}
\keyword{internal}
