% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistar.R
\name{new_multistar}
\alias{new_multistar}
\title{\code{multistar} S3 class}
\usage{
new_multistar(fl = list(), dl = list())
}
\arguments{
\item{fl}{A \code{fact_table} list.}

\item{dl}{A \code{dimension_table} list.}
}
\value{
A \code{multistar} object.
}
\description{
Internal low-level constructor that creates new objects with the correct
structure.
}
\details{
It only distinguishes between general and conformed dimensions, each
dimension has its own data. It can contain multiple fact tables.
}
\keyword{internal}
