% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_update.R
\name{new_record_update}
\alias{new_record_update}
\title{record_update S3 class}
\usage{
new_record_update(dimension, old, new)
}
\value{
A \code{record_update} object.
}
\description{
Internal low-level constructor that creates new objects with the correct
structure.
}
\details{
For a dimension, it relates old record field values to the new values to
replace them.
}
\keyword{internal}
