% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_update_set.R
\name{new_record_update_set}
\alias{new_record_update_set}
\title{record_update_set S3 class}
\usage{
new_record_update_set()
}
\value{
A \code{record_update_set} object.
}
\description{
Internal low-level constructor that creates new objects with the correct
structure.
}
\keyword{internal}
