% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_schema.R
\name{new_star_schema}
\alias{new_star_schema}
\title{star_schema S3 class}
\usage{
new_star_schema(ft = tibble::tibble(), sd = dimensional_model())
}
\arguments{
\item{ft}{A \code{tibble}, implements a flat table.}

\item{sd}{A \code{dimensional_model} object.}
}
\value{
A \code{star_schema} object.
}
\description{
Internal low-level constructor that creates new objects with the correct
structure.
}
\keyword{internal}
