% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{prepare_join}
\alias{prepare_join}
\title{Transform a \code{tibble} to join}
\usage{
prepare_join(tb)
}
\arguments{
\item{tb}{A \code{tibble}.}
}
\value{
A \code{tibble}.
}
\description{
Transform all fields in a \code{tibble} to character type and replace the \code{NA}
with a specific value.
}
\keyword{internal}
