% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_schema_purge_dimensions.R
\name{purge_dimensions}
\alias{purge_dimensions}
\alias{purge_dimensions.star_schema}
\title{Purge dimensions in a star schema}
\usage{
purge_dimensions(st)

\method{purge_dimensions}{star_schema}(st)
}
\arguments{
\item{st}{A \code{star_schema} object.}
}
\value{
A \code{star_schema} object.
}
\description{
Delete instances of dimensions not related to facts.
}
\examples{
library(tidyr)

st <- st_mrs_age \%>\%
  purge_dimensions()

}
\seealso{


Other incremental refresh functions: 
\code{\link{filter_fact_rows}()},
\code{\link{incremental_refresh_constellation}()},
\code{\link{incremental_refresh_star_schema}()}
}
\concept{incremental refresh functions}
