% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellation_purge_dimensions.R
\name{purge_dimensions_constellation}
\alias{purge_dimensions_constellation}
\alias{purge_dimensions_constellation.constellation}
\title{Purge dimensions in a constellation}
\usage{
purge_dimensions_constellation(ct)

\method{purge_dimensions_constellation}{constellation}(ct)
}
\arguments{
\item{ct}{A \code{constellation} object.}
}
\value{
A \code{constellation} object.
}
\description{
Delete instances of dimensions not related to facts in a constellation.
}
\examples{
library(tidyr)

ct <- ct_mrs \%>\%
  purge_dimensions_constellation()

}
\seealso{
Other incremental refresh functions: 
\code{\link{filter_fact_rows}()},
\code{\link{get_star_schema_names}()},
\code{\link{get_star_schema}()},
\code{\link{incremental_refresh_constellation}()},
\code{\link{incremental_refresh_star_schema}()},
\code{\link{purge_dimensions_star_schema}()}
}
\concept{incremental refresh functions}
