% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_schema_purge_dimensions.R
\name{purge_dimensions_star_schema}
\alias{purge_dimensions_star_schema}
\alias{purge_dimensions_star_schema.star_schema}
\title{Purge dimensions}
\usage{
purge_dimensions_star_schema(st)

\method{purge_dimensions_star_schema}{star_schema}(st)
}
\arguments{
\item{st}{A \code{star_schema} object.}
}
\value{
A \code{star_schema} object.
}
\description{
Delete instances of dimensions not related to facts in a star schema.
}
\examples{
library(tidyr)

st <- st_mrs_age \%>\%
  purge_dimensions_star_schema()

}
\seealso{


Other incremental refresh functions: 
\code{\link{filter_fact_rows}()},
\code{\link{incremental_refresh_constellation}()},
\code{\link{incremental_refresh_star_schema}()},
\code{\link{purge_dimensions_constellation}()}
}
\concept{incremental refresh functions}
