% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_update_set.R
\name{record_update_set}
\alias{record_update_set}
\title{\code{record_update_set} S3 class}
\usage{
record_update_set()
}
\value{
A \code{record_update_set} object.
}
\description{
A \code{record_update_set} object is created. Stores updates on dimension records.
}
\details{
Each update is made up of a dimension name, an old value set, and a new value
set.

When the update is applied, all the dimension records that have the
combination of old values are modified with the new values provided.
}
\examples{

updates <- record_update_set()

}
\seealso{
Other data cleaning functions: 
\code{\link{get_conformed_dimension_names}()},
\code{\link{get_conformed_dimension}()},
\code{\link{get_dimension_names}()},
\code{\link{get_dimension}()},
\code{\link{match_records}()},
\code{\link{modify_conformed_dimension_records}()},
\code{\link{modify_dimension_records}()},
\code{\link{update_record}()},
\code{\link{update_selection_general}()},
\code{\link{update_selection}()}
}
\concept{data cleaning functions}
