% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomultistar.R
\name{reference_dimension}
\alias{reference_dimension}
\title{Reference a dimension}
\usage{
reference_dimension(ft, dimension, attributes, conversion = TRUE)
}
\arguments{
\item{ft}{A \code{fact_table} object.}

\item{dimension}{A \code{dimension_table} object.}

\item{attributes}{A vector of attribute names, attributes used to reference the dimension.}

\item{conversion}{A boolean, indicates whether the attributes need to be
transformed.}
}
\value{
A \code{fact_table} object.
}
\description{
Given a dimension, transform the fact table so that the attributes of the
dimension indicated as a parameter, which are in the fact table, are replaced
by the other attributes of the dimension.
}
\details{
It is used to replace a set of attributes in the fact table with the
generated key of the dimension.

If necessary, it is also used for the inverse operation: replace the
generated key with the rest of attributes (dereference a dimension).
}
\keyword{internal}
