% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_query_run_query.R
\name{remove_duplicate_dimension_rows}
\alias{remove_duplicate_dimension_rows}
\title{Remove duplicate dimension rows}
\usage{
remove_duplicate_dimension_rows(dq)
}
\arguments{
\item{dq}{A \code{dimensional_query} object.}
}
\value{
A \code{dimensional_query} object.
}
\description{
After selecting only a few columns of the dimensions, there may be rows with
duplicate values. We eliminate duplicates and adapt facts to the new
dimensions.
}
\keyword{internal}
