% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_schema_rename.R
\name{rename_fact}
\alias{rename_fact}
\alias{rename_fact.star_schema}
\title{Rename fact}
\usage{
rename_fact(st, name)

\method{rename_fact}{star_schema}(st, name)
}
\arguments{
\item{st}{A \code{star_schema} object.}

\item{name}{A string, new name of the fact.}
}
\value{
A \code{star_schema} object.
}
\description{
Set new name for facts.
}
\examples{

st <- st_mrs_age |> rename_fact("age")

}
\seealso{
Other rename functions: 
\code{\link{get_dimension_attribute_names}()},
\code{\link{get_measure_names}()},
\code{\link{rename_dimension}()},
\code{\link{rename_dimension_attributes}()},
\code{\link{rename_measures}()}
}
\concept{rename functions}
