% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_schema_replace_dimension.R
\name{replace_dimension}
\alias{replace_dimension}
\alias{replace_dimension.star_schema}
\title{Replace a star schema dimension}
\usage{
replace_dimension(st, name, dimension)

\method{replace_dimension}{star_schema}(st, name, dimension)
}
\arguments{
\item{st}{A \code{star_schema} object.}

\item{name}{A string, name of the dimension.}

\item{dimension}{A \code{dimension_table} object.}
}
\value{
A \code{star_schema} object.
}
\description{
Replace dimension with another that contains all the instances of the first
and, possibly, some more, in a star schema.
}
\keyword{internal}
