% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_schema_replace_dimension.R
\name{replace_dimension_in_facts}
\alias{replace_dimension_in_facts}
\alias{replace_dimension_in_facts.star_schema}
\title{Replace in facts a star schema dimension}
\usage{
replace_dimension_in_facts(st, name, dimension, set_type_conformed = FALSE)

\method{replace_dimension_in_facts}{star_schema}(st, name, dimension, set_type_conformed = FALSE)
}
\arguments{
\item{st}{A \code{star_schema} object.}

\item{name}{A string, name of the dimension.}

\item{dimension}{A \code{dimension_table} object.}

\item{set_type_conformed}{A boolean.}
}
\value{
A \code{star_schema} object.
}
\description{
This operation can be due to integrating several dimensions in a
constellation or an incremental update of a dimension (indicated with the
boolean parameter). The new dimension replaces in facts the original
dimension, whose name is indicated.
}
\keyword{internal}
