% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_schema_replace_dimension.R
\name{replace_general_dimension_in_facts}
\alias{replace_general_dimension_in_facts}
\title{Replace in facts a star schema general dimension}
\usage{
replace_general_dimension_in_facts(st, name, dimension)
}
\arguments{
\item{st}{A \code{star_schema} object.}

\item{name}{A string, name of the dimension.}

\item{dimension}{A \code{dimension_table} object.}
}
\value{
A \code{star_schema} object.
}
\description{
Replace in facts a star schema general dimension
}
\keyword{internal}
