% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact_table_incremental_refresh.R
\name{replace_records}
\alias{replace_records}
\title{Replace records}
\usage{
replace_records(ft, ft_new, fk)
}
\arguments{
\item{ft}{A \code{fact_table} object.}

\item{ft_new}{A \code{fact_table} object.}

\item{fk}{A vector of foreign key names.}
}
\value{
A \code{fact_table} object.
}
\description{
Replace records with the same primary key.
}
\keyword{internal}
