% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_schema_replace_dimension.R
\name{replace_role_dimension_in_facts}
\alias{replace_role_dimension_in_facts}
\title{Replace in facts a star schema role dimension}
\usage{
replace_role_dimension_in_facts(st, name, dimension, dimension_names)
}
\arguments{
\item{st}{A \code{star_schema} object.}

\item{name}{A string, name of the dimension.}

\item{dimension}{A \code{dimension_table} object.}

\item{dimension_names}{A vector of dimension names.}
}
\value{
A \code{star_schema} object.
}
\description{
Replace in facts a star schema role dimension
}
\keyword{internal}
