% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_table_role_dimension.R
\name{role_dimension}
\alias{role_dimension}
\alias{role_dimension.dimension_table}
\title{Transform a dimension into a role dimension}
\usage{
role_dimension(dimension, role_playing_name)

\method{role_dimension}{dimension_table}(dimension, role_playing_name)
}
\arguments{
\item{dimension}{A \code{dimension_table} object.}

\item{role_playing_name}{A string, name of role-playing dimension.}
}
\value{
A \code{dimension_table} object.
}
\description{
Once the role-playing dimension has been generated, the dimensions from which
it has been defined are transformed into role dimensions. Records are removed
as they are obtained from the role-playing dimension.
}
\keyword{internal}
