% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact_table_incremental_refresh.R
\name{selection_bit_map}
\alias{selection_bit_map}
\title{Generate a record selection bitmap}
\usage{
selection_bit_map(table, values, names)
}
\arguments{
\item{table}{A \code{tibble}, table to select.}

\item{values}{A \code{tibble}, set of values to search.}

\item{names}{A vector of column names to consider.}
}
\value{
A vector of boolean.
}
\description{
Obtain a vector of booleans to select the records in the table that have the
combination of values.
}
\keyword{internal}
