% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_table_attributes.R
\name{set_dimension_name}
\alias{set_dimension_name}
\alias{set_dimension_name.dimension_table}
\title{Set the dimension name}
\usage{
set_dimension_name(dimension, name)

\method{set_dimension_name}{dimension_table}(dimension, name)
}
\arguments{
\item{dimension}{A \code{dimension_table} object.}

\item{name}{A string, name of the dimension.}
}
\value{
A \code{dimension_table} object.
}
\description{
It allows us to define the name of the dimension.
}
\details{
Attributes can be accessed directly but this function has been defined
because it is used from other classes and is thus done in a more controlled
way.
}
\keyword{internal}
