% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_table_attributes.R
\name{set_dimension_type}
\alias{set_dimension_type}
\alias{set_dimension_type.dimension_table}
\title{Set the dimension type}
\usage{
set_dimension_type(dimension, type)

\method{set_dimension_type}{dimension_table}(dimension, type)
}
\arguments{
\item{dimension}{A \code{dimension_table} object.}

\item{type}{A string, type of the dimension.}
}
\value{
A \code{dimension_table} object.
}
\description{
It allows us to define the type of the dimension.
}
\details{
Attributes can be accessed directly but this function has been defined
because it is used from other classes and is thus done in a more controlled
way.
}
\keyword{internal}
