% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_table_attributes.R
\name{set_dimension_type_conformed}
\alias{set_dimension_type_conformed}
\alias{set_dimension_type_conformed.dimension_table}
\title{Set the type of a conformed dimension}
\usage{
set_dimension_type_conformed(dimension)

\method{set_dimension_type_conformed}{dimension_table}(dimension)
}
\arguments{
\item{dimension}{A \code{dimension_table} object.}
}
\value{
A \code{dimension_table} object.
}
\description{
It allows us to define the type of a conformed dimension.
}
\details{
Attributes can be accessed directly but this function has been defined
because it is used from other classes and is thus done in a more controlled
way.
}
\keyword{internal}
