% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact_table.R
\name{set_fact_name}
\alias{set_fact_name}
\alias{set_fact_name.fact_table}
\title{Set fact name}
\usage{
set_fact_name(ft, name)

\method{set_fact_name}{fact_table}(ft, name)
}
\arguments{
\item{ft}{A \code{fact_table} object.}

\item{name}{A string, name of fact.}
}
\value{
A \code{fact_table} object.
}
\description{
It allows us to define the name of facts.
}
\details{
Attributes can be accessed directly but this function has been defined
because it is used from other classes and is thus done in a more controlled
way.
}
\keyword{internal}
