% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_table_attributes.R
\name{set_role_playing_dimension_name}
\alias{set_role_playing_dimension_name}
\alias{set_role_playing_dimension_name.dimension_table}
\title{Set the associated role-playing dimension name}
\usage{
set_role_playing_dimension_name(dimension, name)

\method{set_role_playing_dimension_name}{dimension_table}(dimension, name)
}
\arguments{
\item{dimension}{A \code{dimension_table} object.}

\item{name}{A string, name of role-playing dimension.}
}
\value{
A \code{dimension_table} object.
}
\description{
Each role dimension has the name of the role-playing dimension associated.
This function allows us to set its name.
}
\details{
Attributes can be accessed directly but this function has been defined
because it is used from other classes and is thus done in a more controlled
way.
}
\keyword{internal}
