% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_table_snake_case.R
\name{snake_case_dimension}
\alias{snake_case_dimension}
\alias{snake_case_dimension.dimension_table}
\title{Transform names according to the snake case style in a dimension}
\usage{
snake_case_dimension(dimension)

\method{snake_case_dimension}{dimension_table}(dimension)
}
\arguments{
\item{dimension}{A \code{dimension_table} object.}
}
\value{
A \code{dimension_table} object.
}
\description{
Transform column, attribute and dimension names according to the snake case
style.
}
\keyword{internal}
