% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact_table_snake_case.R
\name{snake_case_fact}
\alias{snake_case_fact}
\alias{snake_case_fact.fact_table}
\title{Transform names according to the snake case style in a fact table}
\usage{
snake_case_fact(ft)

\method{snake_case_fact}{fact_table}(ft)
}
\arguments{
\item{ft}{A \code{fact_table} object.}
}
\value{
A \code{fact_table} object.
}
\description{
Transform foreign keys, measures and fact table names according to the snake
case style.
}
\keyword{internal}
