% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{st_mrs_cause_w10}
\alias{st_mrs_cause_w10}
\title{Star Schema for Mortality Reporting System by Cause for Week 10}
\format{
A \code{star_schema} object.
}
\usage{
st_mrs_cause_w10
}
\description{
Star Schema for the Mortality Reporting System considering the cause
classification data, for week 10 of 1962. It also includes some isolated data
from previous weeks that is supposed to be additional data not considered
before.
}
\examples{
# Defined by:

library(tidyr)

st_mrs_cause_w10 <- star_schema(mrs_cause_w10, dm_mrs_cause) \%>\%
  snake_case() \%>\%
  character_dimensions(
    NA_replacement_value = "Unknown",
    length_integers = list(
      week = 2,
      data_availability_week = 2,
      reception_week = 2
    )
  ) \%>\%
  role_playing_dimension(
    dim_names = c("when", "when_received", "when_available"),
    name = "when_common",
    attributes = c("date", "week", "year")
  )

}
\keyword{datasets}
