% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_schema_as_flat_table.R
\name{star_schema_as_flat_table}
\alias{star_schema_as_flat_table}
\alias{star_schema_as_flat_table.star_schema}
\title{Export a star schema as a flat table}
\usage{
star_schema_as_flat_table(st)

\method{star_schema_as_flat_table}{star_schema}(st)
}
\arguments{
\item{st}{A \code{star_schema} object.}
}
\value{
A \code{tibble}.
}
\description{
Once we have refined the format or content of facts and dimensions, we can
again obtain a flat table, implemented using a \code{tibble}, from a star schema.
}
\examples{
library(tidyr)

ft <- st_mrs_age \%>\%
  star_schema_as_flat_table()

}
\seealso{
Other results export functions: 
\code{\link{constellation_as_multistar}()},
\code{\link{constellation_as_tibble_list}()},
\code{\link{multistar_as_flat_table}()},
\code{\link{star_schema_as_multistar}()},
\code{\link{star_schema_as_tibble_list}()}
}
\concept{results export functions}
