% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_schema_as_multistar.R
\name{star_schema_as_mst}
\alias{star_schema_as_mst}
\alias{star_schema_as_mst.star_schema}
\title{Star schema as multistar export (common)}
\usage{
star_schema_as_mst(st, fl = NULL, dl = NULL, commondim = NULL)

\method{star_schema_as_mst}{star_schema}(st, fl = NULL, dl = NULL, commondim = NULL)
}
\arguments{
\item{st}{A \code{star_schema} object.}

\item{fl}{A list of \code{fact_table} objects.}

\item{dl}{A list of \code{dimension_table} objects.}

\item{commondim}{A list of dimension names already included.}
}
\value{
A \code{multistar} object.
}
\description{
Star schema as multistar export (common)
}
\keyword{internal}
