% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_schema_as_tibble_list.R
\name{star_schema_as_tl}
\alias{star_schema_as_tl}
\alias{star_schema_as_tl.star_schema}
\title{Export a star schema as a \code{tibble} list (common)}
\usage{
star_schema_as_tl(st, tl_prev = NULL, commondim = NULL, include_role_playing)

\method{star_schema_as_tl}{star_schema}(st, tl_prev = NULL, commondim = NULL, include_role_playing)
}
\arguments{
\item{st}{A \code{star_schema} object.}

\item{tl_prev}{A list of \code{tibble} objects.}

\item{commondim}{A list of dimension names already included.}

\item{include_role_playing}{A boolean.}
}
\value{
A \code{tibble} list.
}
\description{
Export a star schema as a \code{tibble} list (common)
}
\keyword{internal}
