% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_table_update_dimension.R
\name{typed_value}
\alias{typed_value}
\title{Transform a value according to its type}
\usage{
typed_value(value, type)
}
\arguments{
\item{value}{A string.}

\item{type}{A string}
}
\value{
A typed value.
}
\description{
Transform a string value according to its given type.
}
\keyword{internal}
