% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_query_run_query.R
\name{unify_facts_by_grain}
\alias{unify_facts_by_grain}
\title{Unify facts by grain}
\usage{
unify_facts_by_grain(dq)
}
\arguments{
\item{dq}{A \code{dimensional_query} object.}
}
\value{
A \code{dimensional_query} object.
}
\description{
Unify facts by grain
}
\keyword{internal}
