% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{union_of_dimensions}
\alias{union_of_dimensions}
\title{Perform union of dimensions}
\usage{
union_of_dimensions(dimensions, name = NULL, type = "role_playing")
}
\arguments{
\item{dimensions}{List of \code{dimension_table} objects.}

\item{name}{A string, name of the dimension.}

\item{type}{A string, type of the dimension.}
}
\value{
A \code{dimension_table} object.
}
\description{
Generates a new dimension from the instances of the dimensions in a list, as
the union of the dimensions.
}
\keyword{internal}
