% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_table_update_dimension.R
\name{update_dimension}
\alias{update_dimension}
\alias{update_dimension.dimension_table}
\title{Apply dimension record update operations to a dimension}
\usage{
update_dimension(dimension, updates)

\method{update_dimension}{dimension_table}(dimension, updates)
}
\arguments{
\item{dimension}{A \code{dimension_table} object.}

\item{updates}{A \code{record_update_set} object.}
}
\value{
A \code{dimension_table} object.
}
\description{
Given a list of dimension record update operations, they are applied on the
\code{dimension_table} object. Update operations must be defined with the set of
functions available for that purpose.
}
\keyword{internal}
