% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{update_dimensions}
\alias{update_dimensions}
\title{Apply update operations to dimensions}
\usage{
update_dimensions(dimensions, updates)
}
\arguments{
\item{dimensions}{List of \code{dimension_table} objects to update.}

\item{updates}{A \code{record_update_set} object.}
}
\value{
List of updated \code{dimension_table} objects.
}
\description{
Apply dimension record update operations to the dimensions in the list.
Returns the list of modified dimensions.
}
\keyword{internal}
