% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_schema_update_facts_with_dimensions.R
\name{update_facts_with_dimensions}
\alias{update_facts_with_dimensions}
\alias{update_facts_with_dimensions.star_schema}
\title{Update facts with a list of modified dimensions}
\usage{
update_facts_with_dimensions(st, dimensions)

\method{update_facts_with_dimensions}{star_schema}(st, dimensions)
}
\arguments{
\item{st}{A \code{star_schema} object.}

\item{dimensions}{A list of \code{dimension_table} objects.}
}
\value{
A \code{star_schema} object.
}
\description{
Update the fact table with the modified dimensions. New dimensions are
generated from the modified ones.
}
\keyword{internal}
