% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_schema_update_facts_with_dimensions.R
\name{update_facts_with_role_dimension}
\alias{update_facts_with_role_dimension}
\title{Update facts with a role dimension}
\usage{
update_facts_with_role_dimension(
  st,
  name,
  old_dimension,
  dimension,
  dimension_names
)
}
\arguments{
\item{st}{A \code{star_schema} object.}

\item{name}{A string, name of the dimension.}

\item{old_dimension}{A \code{dimension_table} object.}

\item{dimension}{A \code{dimension_table} object.}

\item{dimension_names}{A vector of dimension names.}
}
\value{
A \code{star_schema} object.
}
\description{
Update facts with a role dimension
}
\keyword{internal}
