% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usa_nation}
\alias{usa_nation}
\title{USA Nation, 2018}
\format{
A \code{sf}.
}
\source{
\url{https://www2.census.gov/geo/tiger/GENZ2018/shp/cb_2018_us_nation_20m.zip}
}
\usage{
usa_nation
}
\description{
From the original dataset, some fields have been selected and renamed.
}
\examples{

# Read by:
#
# filepath <- "data/cb_2018_us_nation_20m/cb_2018_us_nation_20m.shp"
# usa_nation <- st_read(filepath)  \%>\%
#   dplyr::select(geo_id = GEOID,
#                 name = NAME)

}
\keyword{datasets}
