% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usa_states}
\alias{usa_states}
\title{USA States, 2018}
\format{
A \code{sf}.
}
\source{
\url{https://www2.census.gov/geo/tiger/GENZ2018/shp/cb_2018_us_state_20m.zip}
}
\usage{
usa_states
}
\description{
From the original dataset, some fields have been selected and renamed, and
only includes the Mortality Reporting System states.
}
\examples{

# Read by:
#
# filepath <- "data/cb_2018_us_state_20m/cb_2018_us_state_20m.shp"
# usa_states <- st_read(filepath)  \%>\%
#   dplyr::select(geo_id = GEOID,
#                 state = STUSPS,
#                 state_name = NAME)

}
\keyword{datasets}
