% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{validate_names}
\alias{validate_names}
\title{Validate names}
\usage{
validate_names(defined_names, names, concept = "name", repeated = FALSE)
}
\arguments{
\item{defined_names}{A vector of strings, defined attribute names.}

\item{names}{A vector of strings, new attribute names.}

\item{concept}{A string, treated concept.}

\item{repeated}{A boolean, repeated names allowed.}
}
\value{
A vector of strings, names.
}
\description{
Validate names
}
\keyword{internal}
