% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomultistar_define_geoattribute.R
\name{define_geoattribute}
\alias{define_geoattribute}
\alias{define_geoattribute.geomultistar}
\title{Define geometric attribute}
\usage{
define_geoattribute(
  gms,
  dimension = NULL,
  attribute = NULL,
  from_layer = NULL,
  by = NULL,
  from_attribute = NULL
)

\method{define_geoattribute}{geomultistar}(
  gms,
  dimension = NULL,
  attribute = NULL,
  from_layer = NULL,
  by = NULL,
  from_attribute = NULL
)
}
\arguments{
\item{gms}{A \code{geomultistar} object.}

\item{dimension}{A string, dimension name.}

\item{attribute}{A string, attribute name.}

\item{from_layer}{A \code{sf} object.}

\item{by}{a vector of correspondence of attributes of the dimension with the
\code{sf} layer structure.}

\item{from_attribute}{A string, attribute name.}
}
\value{
A \code{geomultistar} object.
}
\description{
Define geometric attribute
}
\examples{
library(tidyr)

}
\seealso{


Other query functions: 
\code{\link{dimensional_query}()},
\code{\link{filter_dimension}()},
\code{\link{run_query}()},
\code{\link{select_dimension}()},
\code{\link{select_fact}()}
}
\concept{query functions}
