% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ms_mrs}
\alias{ms_mrs}
\title{Multistar for Mortality Reporting System}
\format{
A \code{multistar} object.
}
\usage{
ms_mrs
}
\description{
Multistar for the Mortality Reporting System considering age and cause
classification.
}
\examples{
# Defined by:

library(tidyr)

ms_mrs <- ct_mrs \%>\%
  constellation_as_multistar()

}
\keyword{datasets}
